Refactor(Overview): Restructures the overview functions into seperate jsx files
