+++ frontend/src/features/admin/AdminMenus/Overviews/Overview.jsx
@@ -1,274 +1,11 @@
 import { useState, useEffect } from 'react';
-import { Users, ShoppingBag, DollarSign, RefreshCw, BarChart3, ArrowUpRight, ArrowDownRight, Clock, Package, AlertTriangle} from 'lucide-react';
-import {LineChart, Line, BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer} from 'recharts';
+import { Users, ShoppingBag, DollarSign, RefreshCw, Clock} from 'lucide-react';
 import KPICard from './KPICard/KPICard';
 import StockMonitoring from './StockMonitoring/StockMonitoring';
+import SalesChart from './SalesChart/SalesChart';
+import TopProductsChart from './TopProducts/TopProductsChart';
+import RecentOrdersTable from './RecentOrders/RecentOrders';
 
-
-
-
-const SalesChart = ({ timeRange, loading, error }) => {
-  const salesData = {
-    daily: [
-      { name: 'Mon', revenue: 4000, orders: 24 },
-      { name: 'Tue', revenue: 3000, orders: 18 },
-      { name: 'Wed', revenue: 5000, orders: 32 },
-      { name: 'Thu', revenue: 4500, orders: 28 },
-      { name: 'Fri', revenue: 6000, orders: 38 },
-      { name: 'Sat', revenue: 5500, orders: 35 },
-      { name: 'Sun', revenue: 4200, orders: 26 }
-    ],
-    weekly: [
-      { name: 'Week 1', revenue: 28000, orders: 180 },
-      { name: 'Week 2', revenue: 32000, orders: 210 },
-      { name: 'Week 3', revenue: 35000, orders: 240 },
-      { name: 'Week 4', revenue: 38000, orders: 260 }
-    ],
-    monthly: [
-      { name: 'Jan', revenue: 120000, orders: 800 },
-      { name: 'Feb', revenue: 135000, orders: 920 },
-      { name: 'Mar', revenue: 148000, orders: 1050 },
-      { name: 'Apr', revenue: 142000, orders: 980 },
-      { name: 'May', revenue: 156000, orders: 1120 },
-      { name: 'Jun', revenue: 168000, orders: 1200 }
-    ]
-  };
-
-  if (loading) {
-    return (
-      <div className="h-80 flex items-center justify-center">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-orange-500"></div>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="h-80 flex items-center justify-center">
-        <div className="text-center">
-          <p className="text-red-500 mb-2">Failed to load chart data</p>
-          <button className="text-orange-500 hover:text-orange-600">Try again</button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <ResponsiveContainer width="100%" height={320}>
-      <LineChart data={salesData[timeRange]}>
-        <CartesianGrid strokeDasharray="3 3" stroke="#f0f0f0" />
-        <XAxis 
-          dataKey="name" 
-          stroke="#666"
-          fontSize={12}
-        />
-        <YAxis 
-          stroke="#666"
-          fontSize={12}
-          tickFormatter={(value) => `$${value.toLocaleString()}`}
-        />
-        <Tooltip 
-          contentStyle={{
-            backgroundColor: 'white',
-            border: '1px solid #e5e7eb',
-            borderRadius: '8px',
-            boxShadow: '0 4px 6px -1px rgba(0, 0, 0, 0.1)'
-          }}
-          formatter={(value, name) => [
-            name === 'revenue' ? `$${value.toLocaleString()}` : value,
-            name === 'revenue' ? 'Revenue' : 'Orders'
-          ]}
-        />
-        <Line 
-          type="monotone" 
-          dataKey="revenue" 
-          stroke="#f97316" 
-          strokeWidth={3}
-          dot={{ fill: '#f97316', strokeWidth: 2, r: 4 }}
-          activeDot={{ r: 6, stroke: '#f97316', strokeWidth: 2 }}
-        />
-      </LineChart>
-    </ResponsiveContainer>
-  );
-};
-
-const TopProductsChart = ({ sortBy, loading, error }) => {
-  const [topProducts, setTopProducts] = useState([
-    { name: 'iPhone 15 Pro', units: 245, revenue: 299000, growth: 12.5 },
-    { name: 'MacBook Pro M3', units: 89, revenue: 178000, growth: -3.2 },
-    { name: 'AirPods Pro', units: 156, revenue: 38900, growth: 8.7 },
-    { name: 'iPad Air', units: 78, revenue: 46800, growth: 15.3 },
-    { name: 'Apple Watch', units: 134, revenue: 53600, growth: -1.8 }
-  ]);
-
-  useEffect(() => {
-    const sorted = [...topProducts].sort((a, b) => {
-      return sortBy === 'revenue' ? b.revenue - a.revenue : b.units - a.units;
-    });
-    setTopProducts(sorted);
-  }, [sortBy]);
-
-  if (loading) {
-    return (
-      <div className="h-80 flex items-center justify-center">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-orange-500"></div>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="h-80 flex items-center justify-center">
-        <div className="text-center">
-          <p className="text-red-500 mb-2">Failed to load product data</p>
-          <button className="text-orange-500 hover:text-orange-600">Try again</button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <ResponsiveContainer width="100%" height={320}>
-      <BarChart data={topProducts} layout="horizontal">
-        <CartesianGrid strokeDasharray="3 3" stroke="#f0f0f0" />
-        <XAxis 
-          type="number" 
-          stroke="#666"
-          fontSize={12}
-          tickFormatter={(value) => sortBy === 'revenue' ? `$${(value/1000).toFixed(0)}k` : value}
-        />
-        <YAxis 
-          type="category" 
-          dataKey="name" 
-          stroke="#666"
-          fontSize={12}
-          width={100}
-        />
-        <Tooltip 
-          contentStyle={{
-            backgroundColor: 'white',
-            border: '1px solid #e5e7eb',
-            borderRadius: '8px',
-            boxShadow: '0 4px 6px -1px rgba(0, 0, 0, 0.1)'
-          }}
-          formatter={(value, name) => [
-            sortBy === 'revenue' ? `$${value.toLocaleString()}` : `${value} units`,
-            sortBy === 'revenue' ? 'Revenue' : 'Units Sold'
-          ]}
-        />
-        <Bar 
-          dataKey={sortBy} 
-          fill="#f97316"
-          radius={[0, 4, 4, 0]}
-        />
-      </BarChart>
-    </ResponsiveContainer>
-  );
-};
-
-const RecentOrdersTable = ({ loading }) => {
-  const orders = [
-    {
-      id: '#40755',
-      customer: 'John Doe',
-      amount: '$1,200',
-      status: 'Completed',
-      date: '2024-02-15'
-    },
-    {
-      id: '#40754',
-      customer: 'Jane Smith',
-      amount: '$850',
-      status: 'Processing',
-      date: '2024-02-14'
-    },
-    {
-      id: '#40753',
-      customer: 'Mike Johnson',
-      amount: '$2,300',
-      status: 'Pending',
-      date: '2024-02-14'
-    },
-    {
-      id: '#40752',
-      customer: 'Sarah Williams',
-      amount: '$950',
-      status: 'Completed',
-      date: '2024-02-13'
-    },
-    {
-      id: '#40751',
-      customer: 'Tom Brown',
-      amount: '$1,500',
-      status: 'Cancelled',
-      date: '2024-02-13'
-    }
-  ];
-
-  const getStatusColor = (status) => {
-    switch (status.toLowerCase()) {
-      case 'completed':
-        return 'bg-green-100 text-green-800';
-      case 'processing':
-        return 'bg-blue-100 text-blue-800';
-      case 'pending':
-        return 'bg-yellow-100 text-yellow-800';
-      case 'cancelled':
-        return 'bg-red-100 text-red-800';
-      default:
-        return 'bg-gray-100 text-gray-800';
-    }
-  };
-
-  if (loading) {
-    return (
-      <div className="space-y-4">
-        {[...Array(5)].map((_, index) => (
-          <div key={index} className="animate-pulse">
-            <div className="flex items-center justify-between py-3">
-              <div className="h-4 bg-gray-200 rounded w-20"></div>
-              <div className="h-4 bg-gray-200 rounded w-24"></div>
-              <div className="h-4 bg-gray-200 rounded w-16"></div>
-              <div className="h-6 bg-gray-200 rounded-full w-20"></div>
-              <div className="h-4 bg-gray-200 rounded w-20"></div>
-            </div>
-          </div>
-        ))}
-      </div>
-    );
-  }
-
-  return (
-    <div className="overflow-x-auto">
-      <table className="w-full">
-        <thead>
-          <tr className="border-b">
-            <th className="text-left py-3 px-4 font-medium text-gray-500">Order ID</th>
-            <th className="text-left py-3 px-4 font-medium text-gray-500">Customer</th>
-            <th className="text-left py-3 px-4 font-medium text-gray-500">Amount</th>
-            <th className="text-left py-3 px-4 font-medium text-gray-500">Status</th>
-            <th className="text-left py-3 px-4 font-medium text-gray-500">Date</th>
-          </tr>
-        </thead>
-        <tbody>
-          {orders.map((order) => (
-            <tr key={order.id} className="border-b last:border-b-0 hover:bg-gray-50 transition-colors">
-              <td className="py-3 px-4 font-medium">{order.id}</td>
-              <td className="py-3 px-4">{order.customer}</td>
-              <td className="py-3 px-4 font-medium">{order.amount}</td>
-              <td className="py-3 px-4">
-                <span className={`px-3 py-1 rounded-full text-sm font-medium ${getStatusColor(order.status)}`}>
-                  {order.status}
-                </span>
-              </td>
-              <td className="py-3 px-4 text-gray-500">{order.date}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-};
 
 const Overview = () => {
   const [loading, setLoading] = useState(true);
@@ -375,7 +112,7 @@ const Overview = () => {
         />
       </div>
 
-      {/* Sales Overview and Top Products Side by Side */}
+      {/* Sales Overview and Top Products */}
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
         {/* Sales Overview Chart */}
         <div className="bg-white rounded-lg shadow-sm p-6">
